BUGFIX: expire bfqq if needed in bfq_bfqq_move
If bfqq happens to be in service but empty when bfq_bfqq_move is invoked,
then bfqq is also still marked as busy and bfq_del_bfqq_busy is invoked.
This triggers the failure of the check that bfq_del_bfqq_busy is not
invoked on the in-service queue. In particular, this flow of operations
is imprecise also because it is better to expire and reschedule bfqq
if in service, so as to make sure that it continues to be served only
if it makes sense accordi to its new group memebrship.

This commit fixes these problems by just expring bfqq, as a preliminary
step, if bfqq is in service.

(cherry picked from commit 85b751c5b722a57677a96da1f2fd09854138f4af)
(cherry picked from commit 12a8f6b0c144492bb3ccacf1a9ad43e2bac3c0bd)