diag: Remove BOM in file
This file seems to have a BOM sequence to indicate the file is
UTF-8. There aren't any UTF sequences in the file though, and
this trips up sparse when running on files that include this
header. Remove the BOM so sparse is happy.

Change-Id: Ie25ce60077d1f2b4eb51bf0792635f02d1b33ae7
Signed-off-by: Stephen Boyd <sboyd@codeaurora.org>
Signed-off-by: Mohit Aggarwal <maggarwa@codeaurora.org>
Signed-off-by: Francisco Franco <franciscofranco.1990@gmail.com>
(cherry picked from commit 8c13b477f074923174de26806dd7922d35fbe5ae)